feat: add kyc details section
